medical-rag-system/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 数据库
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# 检索基础能力
numpy>=1.24.0
rank-bm25>=0.2.2

# 文档解析
pymupdf>=1.23.0
python-docx>=1.1.0
python-multipart>=0.0.6

# HTTP客户端
httpx>=0.25.0
aiohttp>=3.9.0

# 工具
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
